# Conflicts:
#	i21Database/dbo/Stored Procedures/uspApiMonthlyRequestUsageStagingTransform.sql
#	i21Database/dbo/Stored Procedures/uspRestApiSchemaTransformItemContract.sql
#	i21Database/dbo/Stored Procedures/uspSTCheckoutPosting.sql
#	i21Database/i21Database.sqlproj

+++ i21Database/dbo/Stored Procedures/uspIPProcessERPStorageLocation.sql
@@ -30,7 +30,7 @@ BEGIN TRY
 		,@dtmDateCreated DATETIME = GETDATE()
 		,@intStorageLocationId INT
 		,@strError NVARCHAR(MAX)
-		,@intTrxSequenceNo INT
+		,@intTrxSequenceNo BIGINT
 	DECLARE @tblICStorageLocation TABLE (
 		strOldDescription NVARCHAR(100)
 		,strNewDescription NVARCHAR(100)
